pyproject.toml
[project]
name = "d-fine"
version = "0.1.0"
description = "D-FINE: Redefine Regression Task of DETRs as Fine-grained Distribution Refinement"
requires-python = ">=3.11"
dependencies = [
    "torch==2.8.0",
    "torchvision==0.23.0",
    "faster-coco-eval==1.7.1",
    "PyYAML==6.0.3",
    "tensorboard==2.20.0",
    "scipy==1.17.0",
    "calflops==0.3.2",
    "transformers==4.57.6",
    "loguru==0.7.3",
    "Pillow==12.1.0",
    "tqdm==4.67.1",
    "matplotlib==3.10.8",
]

[project.optional-dependencies]
dev = []

[tool.uv.sources]
